Updated BlankScreen and package-lock
Added many more dependencies needed for map coding. Implemented map using `react-native-maps` with user location tracking. Implemented a slider to manually adjust the radius of the circle on the map. Added dynamic updates for the circle’s center and radius on timer tick. Styled the app with basic layout and components for a smooth user experience.